floatify

+++ src/main/java/io/github/ageuxo/gloriousgunpowder/data/GunAttribute.java
@@ -8,18 +8,18 @@ import net.minecraft.network.codec.StreamCodec;
 
 import java.util.List;
 //TODO: replace with something even more better
-public record GunAttribute(String stat, double value) {
+public record GunAttribute(String stat, float value) {
     public static final Codec<GunAttribute> CODEC = RecordCodecBuilder.create(instance ->
 
             instance.group(
                     Codec.STRING.fieldOf("stat").forGetter(GunAttribute::stat),
-                    Codec.DOUBLE.fieldOf("value").forGetter(GunAttribute::value)
+                    Codec.FLOAT.fieldOf("value").forGetter(GunAttribute::value)
             ).apply(instance, GunAttribute::new));
     public static final Codec<List<GunAttribute>> LIST_CODEC = CODEC.listOf();
     public static final StreamCodec<RegistryFriendlyByteBuf, GunAttribute> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.STRING_UTF8,
             GunAttribute::stat,
-            ByteBufCodecs.DOUBLE,
+            ByteBufCodecs.FLOAT,
             GunAttribute::value,
             GunAttribute::new
     );
